% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonyms.R
\docType{data}
\name{pseudonyms}
\alias{pseudonyms}
\title{Pseudonyms}
\format{
An object of class \code{character} of length 1123.
}
\usage{
pseudonyms
}
\description{
Some random english words to be used as codewords or pseudonyms.
Useful to identify yourself in a survey results, but not to others.
}
\keyword{datasets}
