% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySurvey-package.R
\docType{package}
\name{shinySurvey-package}
\alias{shinySurvey}
\alias{shinySurvey-package}
\title{shinySurvey: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxheld83/shinySurvey}
  \item Report bugs at \url{https://github.com/maxheld83/shinySurvey/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

}
\keyword{internal}
